test: cover completion diagnostics
