+++ pacquet/crates/cli/src/cli_args/completion.rs
@@ -79,54 +79,4 @@ pub fn shell_from_args(
 }
 
 #[cfg(test)]
-mod tests {
-    use miette::Diagnostic as _;
-
-    use super::{CompletionShell, shell_from_args};
-
-    fn strings(values: &[&str]) -> Vec<String> {
-        values.iter().map(|value| (*value).to_string()).collect()
-    }
-
-    #[test]
-    fn missing_shell_errors_like_pnpm() {
-        let err = shell_from_args(None, &[]).expect_err("missing shell rejected");
-        assert_eq!(err.to_string(), "`pacquet completion` requires a shell name");
-    }
-
-    #[test]
-    fn empty_shell_errors_like_pnpm() {
-        let err = shell_from_args(Some(" \n"), &[]).expect_err("blank shell rejected");
-        assert_eq!(err.to_string(), "`pacquet completion` requires a shell name");
-    }
-
-    #[test]
-    fn unsupported_shell_lists_supported_shells() {
-        let err = shell_from_args(Some("elvish"), &[]).expect_err("unsupported shell rejected");
-        assert_eq!(err.to_string(), "'elvish' is not supported");
-        assert_eq!(
-            err.help().map(|help| help.to_string()),
-            Some("Supported shells are: bash, fish, pwsh, zsh".to_string())
-        );
-    }
-
-    #[test]
-    fn redundant_parameters_are_rejected_before_shell_validation() {
-        let extra = strings(&["fish", "pwsh", "zsh"]);
-        let err = shell_from_args(None, &extra).expect_err("redundant params rejected");
-        assert_eq!(err.to_string(), "The 3 parameters after shell is not necessary");
-    }
-
-    #[test]
-    fn supported_shells_parse_after_trimming() {
-        assert_eq!(shell_from_args(Some(" bash\n"), &[]).expect("bash"), CompletionShell::Bash);
-        assert_eq!(shell_from_args(Some("fish"), &[]).expect("fish"), CompletionShell::Fish);
-        assert_eq!(shell_from_args(Some("pwsh"), &[]).expect("pwsh"), CompletionShell::Pwsh);
-        assert_eq!(shell_from_args(Some("zsh"), &[]).expect("zsh"), CompletionShell::Zsh);
-    }
-
-    #[test]
-    fn pwsh_maps_to_powershell_generator() {
-        assert_eq!(CompletionShell::Pwsh.to_clap_shell(), clap_complete::Shell::PowerShell);
-    }
-}
+mod tests;
